requirements.txt
fastapi[standard]==0.113.0
SQLAlchemy==1.4.49
uvicorn==0.22.0
pytest==7.4.2
pydantic==2.8.0
psycopg2-binary==2.9.6
bcrypt==4.0.1
pydantic==2.8.0
pyjwt==2.10.0
requests==2.32.3
responses==0.25.3
mailjet-rest==1.3.4